tests/downstream-ci-requirements.txt
numpy
pytest-reraise
xarray
earthkit-data @ git+https://github.com/ecmwf/earthkit-data@develop
pint
